go snapshotter: don't fail when there's no file + no snapshots
In order to be able to call snapshotter.Verify regardless of whether any
snapshots were run and not have empty snapshotter files written, I've
added a condition where we skip the comparison when the file doesn't
exist and no snapshots have been run.

Edge-case:
If you renamed a test _and_ made snapshots no-ops in the same commit,
this wouldn't catch the issue. I'm ok with this.