snapshotter: don't emit `null` snapshot files
Previously if you initalized snapshotter in a test & invoked Verify()
at the end, if no snapshots were recorded, a `null` file would be
output. This was clutter for folks to avoid adding to git.

Here we simply don't write the file if we see no snapshots recorded during the test.

For testing, there already exists a test verifying no snapshots work, and it seems that nobody
ever ran the test with -rewriteSnapshots, so there isn't a diff removing testdata, but it covers
our case here, and I tested it locally.