snapshotter: rewriteWithFailOnDiff
Prior to this commit, there were only two options:
rewrite or verify mode.

If you are in rewrite mode, we rewrite the snapshots and the unit test
will never fail on the Verify() call (unless something goes wrong with
the mechanics of writing the snapshots file).

Else, if you are in verify mode (the default mode), we will compare the
snapshots with the actual values and fail the unit test if there is any
mismatch.

This commit introduces a third option: we first verify, and then we
rewrite the snapshot. This might be a very specialized use-case, but at
Samsara, we have developer tooling around making it easier to patch
these snapshot diffs, which involve the snapshot diffs being generated
(so we need rewrite mode) and pushed to our CI artifacts bucket for
later application, but we also want the unit test to fail so
that our CI flake detection mechanisms can properly flag the test as
flaky.