oops: fix test suite
This fixes the existing test suite for the oops library. The pre-existing tests were a little
brittle, relying on the checkout dir when comparing expected stack traces to captured ones.
Also, featured a lot of hard-coded comparisons to line numbers in expected stack traces

The test suite should now run ok regardless of the parent folder structure of the git checkout, also
it should be tolerant of differening dependencies / go versions that may re-order LOC. Also, when
adding or removing lines in the test file itself, there should no longer be a need to update the
hard-coded assertions of line numbers.